fix(playerbot): 64-bit yang in log formats crashed the core on a Teleporter refusal; consumables and bonus stones by subtype
- travel.h and twelve other log lines passed GetGold() - long long on the
  2.x line - to %d. The Teleporter refusal had gold in the middle of the
  list with reason=%s after it, so every refusal read the map number as a
  char pointer and killed the core; the line never reached a log
  (kimakatsu, 13 September, diagnosed from the stack trace). Cast to
  long long and %lld everywhere gold is logged.
- consumables/economy/bonus/town: a booster is USE_AFFECT value0 510 or
  USE_ABILITY_UP, a bonus stone its subtype whatever its vnum, an exp
  elixir is drunk on sight, the Metin detector is counter goods, and the
  junk rule exempts all of them - the ItemShop copies (76xxx) carry no
  ANTI_SELL and a bot handed them from the panel vendored the lot
  (Pasywny).

Both engines syntax-checked clean, no format warnings.

+++ linux-port/overlays/playerbot/src/game/src/playerbot_market.h
@@ -358,10 +358,10 @@ namespace
 		RememberPlayerBotSale(pick.dwVnum, pick.bRefine,
 				(DWORD)paid / std::max<DWORD>(1, pick.wCount), get_dword_time(),
 				pick.dwSkillVnum);
-		sys_log(0, "PLAYERBOT_MARKET: bought pid=%u name=%s from=%s slot=%u vnum=%u refine=%u count=%u asked=%u paid=%d gold=%d",
+		sys_log(0, "PLAYERBOT_MARKET: bought pid=%u name=%s from=%s slot=%u vnum=%u refine=%u count=%u asked=%u paid=%lld gold=%lld",
 				ch->GetPlayerID(), ch->GetName(), pick.keeper->GetName(),
 				(unsigned int)pick.bSlot, pick.dwVnum, (unsigned int)pick.bRefine,
-				(unsigned int)pick.wCount, pick.dwPrice, paid, ch->GetGold());
+				(unsigned int)pick.wCount, pick.dwPrice, (long long)paid, (long long)ch->GetGold());
 		return true;
 	}
